requirements.txt
appnope==0.1.3
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
black==22.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
dj-rest-auth==2.2.5
Django==4.0.4
django-allauth==0.50.0
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-filter==22.1
django-stubs==1.11.0
django-stubs-ext==0.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djangorestframework-stubs==1.6.0
entrypoints==0.4
executing==0.8.3
Faker==13.15.1
filelock==3.6.0
httpie==3.2.1
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.13.0
ipython==8.3.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
jupyter-client==7.3.0
jupyter-core==4.10.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
multidict==6.0.2
mypy==0.950
mypy-extensions==0.4.3
nest-asyncio==1.5.5
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.5.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.8
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
pyzmq==22.3.0
requests==2.28.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.4.4
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.1
types-Markdown==3.3.28
types-pytz==2021.3.8
types-PyYAML==6.0.8
types-requests==2.27.30
types-urllib3==1.26.15
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.13.1
wcwidth==0.2.5
zipp==3.8.1
